fixing testcases
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/branches/distributed@12955 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ exp/include/Galois/Runtime/DistSupport.h
@@ -147,7 +147,7 @@ public:
   }
 
   void dump() {
-    printf("[%x,%lx]", owner, ptr);
+    printf("[%x,%lx]", owner, (size_t)ptr);
   }
 };
